mwindow: set limit on number of open files
Doing many network operations can lead to lots of small packfiles.
The existing mwindow limit was only the total size of mmap'd files,
not on their number, so having lots of small packfiles could exhaust
the allowed number of open files, particularly on macOS, where
the default ulimit is very low.

Add a limit to the number of files as well as their total size,
and set the default limit low enough that even macOS users should not
hit it during normal use.

Fixes #2758